+++ nautilus_trader/backtest/engine.pyx
@@ -1207,16 +1207,21 @@ cdef class BacktestEngine:
                 for exchange in self._venues.values():
                     exchange.process(ts_event_init)
 
+    def _get_log_color_code(self):
+        return "\033[36m" if self._log.is_colored else ""
+
     def _log_pre_run(self):
         log_memory(self._log)
 
+        cdef str color = self._get_log_color_code()
+
         for exchange in self._venues.values():
             account = exchange.exec_client.get_account()
-            self._log.info("\033[36m=================================================================")
-            self._log.info(f"\033[36m SimulatedVenue {exchange.id}")
-            self._log.info("\033[36m=================================================================")
+            self._log.info(f"{color}=================================================================")
+            self._log.info(f"{color} SimulatedVenue {exchange.id}")
+            self._log.info(f"{color}=================================================================")
             self._log.info(f"{repr(account)}")
-            self._log.info("\033[36m-----------------------------------------------------------------")
+            self._log.info(f"{color}-----------------------------------------------------------------")
             self._log.info(f"Balances starting:")
             if exchange.is_frozen_account:
                 self._log.warning(f"ACCOUNT FROZEN")
@@ -1225,16 +1230,18 @@ cdef class BacktestEngine:
                     self._log.info(b.to_str())
 
     def _log_run(self, start: pd.Timestamp, end: pd.Timestamp):
-        self._log.info("\033[36m=================================================================")
-        self._log.info("\033[36m BACKTEST RUN")
-        self._log.info("\033[36m=================================================================")
+        cdef str color = self._get_log_color_code()
+
+        self._log.info(f"{color}=================================================================")
+        self._log.info(f"{color} BACKTEST RUN")
+        self._log.info(f"{color}=================================================================")
         self._log.info(f"Run config ID:  {self._run_config_id}")
         self._log.info(f"Run ID:         {self._run_id}")
         self._log.info(f"Run started:    {self._run_started}")
         self._log.info(f"Backtest start: {self._backtest_start}")
         self._log.info(f"Batch start:    {start}")
         self._log.info(f"Batch end:      {end}")
-        self._log.info("\033[36m-----------------------------------------------------------------")
+        self._log.info(f"{color}-----------------------------------------------------------------")
 
     def _log_post_run(self):
         if self._run_finished and self._run_started:
@@ -1247,9 +1254,11 @@ cdef class BacktestEngine:
         else:
             backtest_range = None
 
-        self._log.info("\033[36m=================================================================")
-        self._log.info("\033[36m BACKTEST POST-RUN")
-        self._log.info("\033[36m=================================================================")
+        cdef str color = self._get_log_color_code()
+
+        self._log.info(f"{color}=================================================================")
+        self._log.info(f"{color} BACKTEST POST-RUN")
+        self._log.info(f"{color}=================================================================")
         self._log.info(f"Run config ID:  {self._run_config_id}")
         self._log.info(f"Run ID:         {self._run_id}")
         self._log.info(f"Run started:    {self._run_started}")
@@ -1277,11 +1286,11 @@ cdef class BacktestEngine:
             set venue_currencies
         for venue in self._venues.values():
             account = venue.exec_client.get_account()
-            self._log.info("\033[36m=================================================================")
-            self._log.info(f"\033[36m SimulatedVenue {venue.id}")
-            self._log.info("\033[36m=================================================================")
+            self._log.info(f"{color}=================================================================")
+            self._log.info(f"{color} SimulatedVenue {venue.id}")
+            self._log.info(f"{color}=================================================================")
             self._log.info(f"{repr(account)}")
-            self._log.info("\033[36m-----------------------------------------------------------------")
+            self._log.info(f"{color}-----------------------------------------------------------------")
             unrealized_pnls: Optional[dict[Currency, Money]] = None
             if venue.is_frozen_account:
                 self._log.warning(f"ACCOUNT FROZEN")
@@ -1291,15 +1300,15 @@ cdef class BacktestEngine:
                 self._log.info(f"Balances starting:")
                 for b in account.starting_balances().values():
                     self._log.info(b.to_str())
-                self._log.info("\033[36m-----------------------------------------------------------------")
+                self._log.info(f"{color}-----------------------------------------------------------------")
                 self._log.info(f"Balances ending:")
                 for b in account.balances_total().values():
                     self._log.info(b.to_str())
-                self._log.info("\033[36m-----------------------------------------------------------------")
+                self._log.info(f"{color}-----------------------------------------------------------------")
                 self._log.info(f"Commissions:")
                 for c in account.commissions().values():
                     self._log.info(Money(-c.as_double(), c.currency).to_str())  # Display commission as negative
-                self._log.info("\033[36m-----------------------------------------------------------------")
+                self._log.info(f"{color}-----------------------------------------------------------------")
                 self._log.info(f"Unrealized PnLs (included in totals):")
                 unrealized_pnls = self.portfolio.unrealized_pnls(Venue(venue.id.value))
                 if not unrealized_pnls:
@@ -1312,9 +1321,9 @@ cdef class BacktestEngine:
             for module in venue.modules:
                 module.log_diagnostics(self._log)
 
-            self._log.info("\033[36m=================================================================")
-            self._log.info("\033[36m PORTFOLIO PERFORMANCE")
-            self._log.info("\033[36m=================================================================")
+            self._log.info(f"{color}=================================================================")
+            self._log.info(f"{color} PORTFOLIO PERFORMANCE")
+            self._log.info(f"{color}=================================================================")
 
             # Collect all positions and currencies for venue
             venue_positions = []
@@ -1332,23 +1341,23 @@ cdef class BacktestEngine:
             # Present PnL performance stats per asset
             for currency in sorted(list(venue_currencies), key=lambda x: x.code):
                 self._log.info(f" PnL Statistics ({str(currency)})")
-                self._log.info("\033[36m-----------------------------------------------------------------")
+                self._log.info(f"{color}-----------------------------------------------------------------")
                 unrealized_pnl = unrealized_pnls.get(currency) if unrealized_pnls else None
                 for stat in self._kernel.portfolio.analyzer.get_stats_pnls_formatted(currency, unrealized_pnl):
                     self._log.info(stat)
-                self._log.info("\033[36m-----------------------------------------------------------------")
+                self._log.info(f"{color}-----------------------------------------------------------------")
 
             self._log.info(" Returns Statistics")
-            self._log.info("\033[36m-----------------------------------------------------------------")
+            self._log.info(f"{color}-----------------------------------------------------------------")
             for stat in self._kernel.portfolio.analyzer.get_stats_returns_formatted():
                 self._log.info(stat)
-            self._log.info("\033[36m-----------------------------------------------------------------")
+            self._log.info(f"{color}-----------------------------------------------------------------")
 
             self._log.info(" General Statistics")
-            self._log.info("\033[36m-----------------------------------------------------------------")
+            self._log.info(f"{color}-----------------------------------------------------------------")
             for stat in self._kernel.portfolio.analyzer.get_stats_general_formatted():
                 self._log.info(stat)
-            self._log.info("\033[36m-----------------------------------------------------------------")
+            self._log.info(f"{color}-----------------------------------------------------------------")
 
     def _add_data_client_if_not_exists(self, ClientId client_id) -> None:
         if client_id not in self._kernel.data_engine.registered_clients: